+++ cmp/grid/impl/InitPersist.ts
@@ -25,6 +25,7 @@ export function initPersist(
     gridModel: GridModel,
     {
         persistColumns = true,
+        persistColumnGroups = true,
         persistGrouping = true,
         persistSort = true,
         persistExpandToLevel = true,
@@ -60,6 +61,21 @@ export function initPersist(
                             gridModel.autosizeAsync({columns});
                         }
                     })
+            },
+            owner: gridModel
+        });
+    }
+
+    if (persistColumnGroups) {
+        PersistenceProvider.create({
+            persistOptions: persistOptions(
+                {path: `${path}.columnGroups`},
+                rootPersistWith,
+                isObject(persistColumnGroups) ? persistColumnGroups : null
+            ),
+            target: {
+                getPersistableState: () => new PersistableState(gridModel.columnGroupState),
+                setPersistableState: ({value}) => gridModel.setColumnGroupState(value)
             },
             owner: gridModel
         });